requirements.txt
asgiref==3.7.2
astroid==3.0.1
autopep8==2.0.4
bcrypt==4.0.1
dill==0.3.7
Django==4.2.6
django-filter==23.3
django-redis==5.4.0
djangorestframework==3.14.0
isort==5.12.0
Markdown==3.5
mccabe==0.7.0
mysqlclient==2.1.1
platformdirs==3.11.0
pycodestyle==2.11.1
PyJWT==2.8.0
pylint==3.0.1
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
sqlparse==0.4.4
tomlkit==0.12.1
